Pointer now shows up with the DRM backend
It turns out that the input devices are added in a different order
with different backends. The reason why the DRM backend wasn't working
was that the outputs were added before the seats. Since the order
shouldn't matter, each new seat needs to look at the available outputs
when they are created.