Don't hardcode c library path
Not only will this not run on other users' machines, it makes it
impossible to install without a rebuild.

This now requires LD_LIBRARY_PATH to be set to the lib folders in the
build directory in order for mahogany to run in a dev environment. Add
a new `run` target in the makefile so it doesn't need to be typed
every time.