Store the log level as an integer, not a symbol
This allows us to effectively inline the calls to
`get-print-data` (which we have renamed to `get-log-level-data`). SBCL
was also having trouble with optimizing the comparison for the if
statement, and this should fix that as well.

To still have the debug level set via symbols, hide the *log-level*
variable and translate between the symbol and number when the value is
being set / queried.