Don't completely inline log functions with macros
Do the work of `log-string` and `log-stream` in acual
functions. This vastly improves the readability of the disassembled
output of the functions that use them, as the branching statements and
other code not longer appears in them. As a side benefit, it also
means that the log level doesn't need to be specified at compile time.
+ To prevent log-string from evaluating it's arguments, build a lambda
  and pass it to log-stream.

In order to continue avoiding the meta data lookup, add a compiler
macro for `log-stream` that will do the lookup if the log levels are
specified at compile time. This makes the debug output slightly worse
than just the raw functions, but it's still better than it was.