hrt: don't keep list of available outputs
Keeping a list of the available outputs just duplicates work that we
have to do in the frontend anyways; the only reason why hrt should
keep it is if we do things with the list of outputs without input
from the front end. We might do that eventually for things like output
configuration, but we aren't anywhere near that yet.