Add circular list to handle minimized views
The ring-list object is a circular list that you can add, remove, and
swap items in. There is some confusion over where items should be
added and what moving forward and backward over the list should do,
but it represents the basic principle of what we need. The interface
is the important part, and we can deal with behavior if this doesn't
work out.
+ Add extremely basic tests