Make build system work on more platforms
The place where .so files are placed are platform dependent; for
example, Fedora places them in $PREFIX/lib64 or $PREFIX/lib based on
the archecture size. Ubuntu uses somethig like
$PREFIX/lib/x86_64-linux-gnu/.

To fix this, set LIBDIR to `lib` when configuring heart so that it
goes into the same place every time.