lisp: Create a tree-parent class so more code can be shared
This class represents any node in the frame tree that has children.

This is a bit of an awkward setup, as there's a class-precedence issue
between the tree-frame type and the other parent types; we have to be
careful not to override the frame methods that the tree-frame class
needs to use.