Round new frame width/height when splitting
Applying this commit will address frame splitting issues where the resultant
frame height/width is a non-integer number, causing an error to be
signalled. This issue occurs because the C heart expect an integer, however CL
does not do integer division by default, so the resultant frame width/height may
not be an integer.