fix frame width calculation for old frame in split
Applying this commit will fix an issue with calculating the new width of the old
frame, whereby the old frame was assumed to have the same new width as the new
frame being created. This is correct if and only if the original width is a
number evenly divisible by the ratios divisor. E.g. when splitting in half with
1405 this would calculate both the old frames new width and the width of the new
frame as 703, which when taken together gives a total width of 1406. This fix
makes it so the widths become 702 and 703, giving the correct total size.